requirements.txt
# ── Core ──────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# ── Data ──────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# ── ML ────────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0

# ── Database ──────────────────────────────────────────────────
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# ── Scheduling / Async ────────────────────────────────────────
apscheduler>=3.10.0
asyncio-mqtt>=0.16.0

# ── Dashboard API ─────────────────────────────────────────────
websockets>=12.0

# ── Utils ─────────────────────────────────────────────────────
rich>=13.7.0
tabulate>=0.9.0
python-dateutil>=2.9.0
pytz>=2024.1
